enabled back unit test and seperated command interpeter to pass compilation
