+++ .circleci/config.yml
@@ -12,7 +12,7 @@ jobs:
           command: export TERM=xterm; ./etc/bsys/bin/makeUnitTests.sh
       - run:
           name: testing_unit
-          command: echo 0
+          command: export CI_BUILD="CCI_UNIT_TEST$CIRCLE_BUILD_NUM"; export CI_CHECK=1; export TERM=xterm; xvfb-run ./sqa/bin/runUnitTests.sh
       - run:
           name: testing_integrations
           command: export CI_BUILD="CCI_INTG_TEST$CIRCLE_BUILD_NUM"; export CI_CHECK=1; export TERM=xterm; export PAINTER_QA_XVFB=1; ./sqa/bin/runTestsParallel.sh
