+++ etc/bsys/cmake_glavni.cm
@@ -18,8 +18,8 @@ if (WIN32)
 	set(QT_INCLUDE_DIR ${QTDIR}/include)
 #linux
 else()
-	set(LD_AUX "dl")
-	set(CMAKE_CXX_FLAGS "-I/usr/include/python3.10 -I/usr/include/python3.10  -Wno-unused-result -Wsign-compare -g      -fstack-protector-strong -Wformat -Werror=format-security  -DNDEBUG -g -fwrapv -O2 -Wall -DNO_QT -DOS_LINUX -g  -lpython3.10 -fPIC -Wunused-parameter -pipe -std=gnu++17 -frtti -Wall -Wextra -fexceptions -pthread -Wno-unused-function -Wno-unused-variable")
+	set(LD_AUX "dl -lpython3.10")
+    set(CMAKE_CXX_FLAGS "-DNO_QT -DOS_LINUX -g -fPIC -Wunused-parameter -pipe -std=gnu++17 -frtti -Wall -Wextra -fexceptions -pthread -Wno-unused-function -Wno-unused-variable")
     set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin/linux/)
     #linux64
     if (EXISTS "/usr/include/x86_64-linux-gnu/qt5/")
