+++ src/commands/test/CMakeLists.txt
@@ -5,7 +5,7 @@ include (../../../etc/bsys/cmake_glavni.cm)
 include_directories(../)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ../../../unit_test_bin)
 
-make_ut(TESTEE shape_creation_directive_commands DEPS ../../core/iobject_pool.cpp ../../core/design.cpp )
+#make_ut(TESTEE shape_creation_directive_commands DEPS ../../core/iobject_pool.cpp ../../core/design.cpp )
 make_ut(TESTEE load_save_commands DEPS ../../core/iobject_pool.cpp ../../core/design.cpp )
-make_ut(TESTEE delete_command DEPS ../../core/iobject_pool.cpp ../../core/design.cpp )
+#make_ut(TESTEE delete_command DEPS ../../core/iobject_pool.cpp ../../core/design.cpp )
